+++ lib/src/utils.dart
@@ -36,6 +36,7 @@ Future<ProcessResult> runProc(
   String workingDirectory,
   Map<String, String> environment,
   Duration timeout,
+  bool deduplicate = false,
 }) async {
   log.info('Running `${([executable]..addAll(arguments)).join(' ')}`...');
   var process = await Process.start(executable, arguments,
@@ -65,6 +66,10 @@ Future<ProcessResult> runProc(
   var items = await Future.wait(<Future<Object>>[
     process.exitCode,
     byteStreamSplit(process.stdout).forEach((outLine) {
+      // TODO: Remove deduplication when https://github.com/dart-lang/sdk/issues/36062 gets fixed
+      if (deduplicate && stdoutLines.contains(outLine)) {
+        return;
+      }
       stdoutLines.add(outLine);
       // Uncomment to debug long execution
       // stderr.writeln(outLine);
@@ -73,6 +78,10 @@ Future<ProcessResult> runProc(
       }
     }),
     byteStreamSplit(process.stderr).forEach((errLine) {
+      // TODO: Remove deduplication when https://github.com/dart-lang/sdk/issues/36062 gets fixed
+      if (deduplicate && stderrLines.contains(errLine)) {
+        return;
+      }
       stderrLines.add(errLine);
       // Uncomment to debug long execution
       // stderr.writeln(errLine);